refactor: Move config and router modules to utils
The utils package contains modules which often do not change when new
features are added and which are written to simplify the code.

The misc package contains modules that change when new features are
added. For example, if we want to introduce a payer entity, we have to
change the models module, which is in misc.